network: Refactor network tests.
Split network_test.go into separate test files.
